+++ app/src/main/java/com/fibelatti/photowidget/home/SettingsScreen.kt
@@ -33,7 +33,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.layout.onLayoutRectChanged
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalInspectionMode
@@ -218,8 +218,8 @@ private fun SettingsScreen(
         SettingsFooter(
             onViewLicensesClick = onViewLicensesClick,
             modifier = Modifier
-                .onGloballyPositioned {
-                    footerHeight = with(localDensity) { it.size.height.toDp() }
+                .onLayoutRectChanged {
+                    footerHeight = with(localDensity) { it.height.toDp() }
                 }
                 .background(
                     brush = Brush.verticalGradient(